test(api): add new file for api tests
feature/225415-AddUserToFeatureToggle
